Adds simple locking lib
