+++ catalogue_pipeline/matcher/src/main/scala/uk/ac/wellcome/platform/matcher/lockable/Locked.scala
@@ -1,0 +1,7 @@
+package uk.ac.wellcome.platform.matcher.lockable
+
+case class Locked[T](t: T) {
+  def unlock(implicit lockable: Lockable[T]): Either[UnlockFailure, T]= {
+    lockable.unlock(this)
+  }
+}
