+++ catalogue_pipeline/matcher/src/main/scala/uk/ac/wellcome/platform/matcher/lockable/LockingFailures.scala
@@ -1,0 +1,6 @@
+package uk.ac.wellcome.platform.matcher.lockable
+
+sealed trait LockingFailures
+
+case class LockFailure(value: String) extends LockingFailures
+case class UnlockFailure(value: String) extends LockingFailures
