+++ catalogue_pipeline/matcher/src/test/scala/uk/ac/wellcome/platform/matcher/fixtures/LocalLockTableDynamoDb.scala
@@ -1,0 +1,32 @@
+package uk.ac.wellcome.platform.matcher.fixtures
+
+import com.amazonaws.services.dynamodbv2.model._
+import com.amazonaws.services.dynamodbv2.util.TableUtils.waitUntilActive
+import uk.ac.wellcome.storage.test.fixtures.LocalDynamoDb
+
+trait LocalLockTableDynamoDb
+  extends LocalDynamoDb {
+  override def createTable(table: LocalDynamoDb.Table): LocalDynamoDb.Table = {
+    dynamoDbClient.createTable(
+      new CreateTableRequest()
+        .withTableName(table.name)
+        .withKeySchema(new KeySchemaElement()
+          .withAttributeName("id")
+          .withKeyType(KeyType.HASH))
+        .withAttributeDefinitions(
+          new AttributeDefinition()
+            .withAttributeName("id")
+            .withAttributeType("S")
+        )
+        .withProvisionedThroughput(new ProvisionedThroughput()
+          .withReadCapacityUnits(1L)
+          .withWriteCapacityUnits(1L))
+    )
+
+    eventually {
+      waitUntilActive(dynamoDbClient, table.name)
+    }
+
+    table
+  }
+}
