+++ catalogue_pipeline/matcher/src/test/scala/uk/ac/wellcome/platform/matcher/fixtures/MatcherFixtures.scala
@@ -32,6 +32,7 @@ trait MatcherFixtures
     with SQS
     with SNS
     with LocalLinkedWorkDynamoDb
+    with LocalLockTableDynamoDb
     with MetricsSenderFixture
     with S3 {
